+++ packages/client/src/subscribe/Resends.ts
@@ -3,9 +3,7 @@
  */
 import { inject, Lifecycle, scoped, delay } from 'tsyringe'
 import { MessageRef, StreamPartID, StreamPartIDUtils, StreamMessage } from 'streamr-client-protocol'
-
-import { createSubscribePipeline } from './SubscribePipeline'
-
+import { createSubscribePipeline } from './subscribePipeline'
 import { StorageNodeRegistry } from '../registry/StorageNodeRegistry'
 import { StreamIDBuilder } from '../StreamIDBuilder'
 import { StreamDefinition } from '../types'
